app/requirements.txt
# Flask web framework
flask==3.0.0
gunicorn==21.2.0

# Azure Blob Storage SDK
azure-storage-blob==12.19.0

# YAML parsing
pyyaml==6.0.1

# Cache with TTL
cachetools==5.3.2

# Environment variables
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Linting
flake8==6.1.0
